feat(core): Ability to create snapshots from existing snapshots (#16597)
