+++ cli/js/40_testing.js
@@ -1397,12 +1397,16 @@
     return testFn;
   }
 
-  window.__bootstrap.testing = {
-    bench,
-    enableBench,
-    enableTest,
-    runBenchmarks,
-    runTests,
-    test,
+  window.__bootstrap.internals = {
+    ...window.__bootstrap.internals ?? {},
+    testing: {
+      runTests,
+      runBenchmarks,
+      enableTest,
+      enableBench,
+    },
   };
+
+  window.__bootstrap.denoNs.bench = bench;
+  window.__bootstrap.denoNs.test = test;
 })(this);
